requirements.txt
fastapi
uvicorn
langchain
flask
openai
python-dotenv
pinecone-client
transformers
git+https://github.com/supabase-community/supabase-py.git
torch
requests
